Revert "fix: move install buildkit to up cmd (#221)"
This reverts commit 08195900a8339be72807d1091156f45a82e8a595.
